requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-debug-toolbar==4.3.0
flake8==7.0.0
idna==3.6
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.4.0
typing_extensions==4.10.0
urllib3==2.2.1